split listing contract draft into two componenents
